requirements.txt
autopep8==1.5.6
argon2-cffi==20.1.0
arrow==1.0.3
aiohttp==3.7.4
asyncpg==0.22.0
python-dotenv==0.15.0
quart==0.14.1
quart-jwt-extended==0.0.3